requirements.txt
streamlit
pdfminer.six
pandas
scikit-learn
matplotlib
numpy